msgs: inline ccs module
